+++ examples/data_types.js
@@ -1,6 +1,6 @@
 const hive = require('../');
 const TCLIService_types = hive.thrift.TCLIService_types;
-const connection = require('./connections/kerberosHttpSsl');
+const connection = require('./connections/hdinsight');
 
 const utils = new hive.HiveUtils(
     TCLIService_types
@@ -12,14 +12,13 @@ connection().then(client => {
     });
     
     return client.openSession({
-        client_protocol: TCLIService_types.TProtocolVersion.HIVE_CLI_SERVICE_PROTOCOL_V10
-    }).then((session) => {
-        return testPrimitiveTypes(session)
-            .then(() => testComplexTypes(session))
-            .then(() => testIntervals(session))
-            .then(() => {
-                return session.close();
-            });
+        client_protocol: TCLIService_types.TProtocolVersion.HIVE_CLI_SERVICE_PROTOCOL_V8
+    }).then(async (session) => {
+        await testPrimitiveTypes(session);
+        await testComplexTypes(session); 
+        await testIntervals(session);
+        
+        return await session.close();
     }).then(status => {
         console.log(status.success());
 
@@ -30,25 +29,29 @@ connection().then(client => {
     });
 });
 
-const testPrimitiveTypes = (session) => {
-    return execute(session, `
-        CREATE TABLE primitiveTypes (
-            bool boolean,
-            tiny_int tinyint,
-            small_int smallint,
-            int_type int,
-            big_int bigint,
-            flt float,
-            dbl double,
-            dec decimal(3,2),
-            str string,
-            ts timestamp,
-            bin binary,
-            chr char(10),
-            vchr varchar(10),
-            dat date
-        )`)
-        .then(() => execute(session, `insert into primitiveTypes (
+const testPrimitiveTypes = async (session) => {
+    try {
+        console.log('[info] create primitiveTypes');
+        await execute(session, `
+            CREATE TABLE IF NOT EXISTS primitiveTypes (
+                bool boolean,
+                tiny_int tinyint,
+                small_int smallint,
+                int_type int,
+                big_int bigint,
+                flt float,
+                dbl double,
+                dec decimal(3,2),
+                str string,
+                ts timestamp,
+                bin binary,
+                chr char(10),
+                vchr varchar(10),
+                dat date
+            )`
+        );
+        console.log('[info] insert into primitiveTypes');
+        await execute(session, `insert into primitiveTypes (
             bool,
             tiny_int,
             small_int,
@@ -78,112 +81,108 @@ const testPrimitiveTypes = (session) => {
             'a',
             'b',
             '2014-01-17'
-        )`))
-        .then(() => {
-            return execute(session, 'select * from primitiveTypes');
-        })
-        .then(console.log)
-        .then(
-            () =>  execute(session, 'drop table primitiveTypes'),
-            (error) => {
-                console.error(error);
-                execute(session, 'drop table primitiveTypes')
-            }
-        );
+        )`);
+        console.log('[info] fetch primitiveTypes');
+        const result = await execute(session, 'select * from primitiveTypes');
+        
+        console.log(result);
+    } finally {
+        await execute(session, 'drop table primitiveTypes');
+    }
 };
 
-const testIntervals = (session) => {
-    return execute(session, `
-        CREATE TABLE intervalTypes AS
-            SELECT INTERVAL '1' day AS day_interval, 
-            INTERVAL '1' month AS month_interval
-    `)
-    .then(() => execute(session, `desc intervalTypes`))
-    .then(console.log)
-    .then(() => execute(session, 'select * from intervalTypes'))
-    .then(console.log)
-    .then(
-        () =>  execute(session, 'drop table intervalTypes'),
-        () =>  execute(session, 'drop table intervalTypes')
-    );
+const testIntervals = async (session) => {
+    try {
+        console.log('[info] create intervalTypes');
+        await execute(session, `
+            CREATE TABLE IF NOT EXISTS intervalTypes AS
+                SELECT INTERVAL '1' day AS day_interval, 
+                INTERVAL '1' month AS month_interval
+        `);
+        console.log('[info] describe intervalTypes');
+        console.log(await execute(session, `desc intervalTypes`));
+        console.log('[info] fetch intervalTypes');
+        console.log(await execute(session, 'select * from intervalTypes'));
+    } finally {
+        await execute(session, 'drop table intervalTypes');
+    }
 };
 
-const testComplexTypes = (session) => {
-    return execute(session, `create table dummy( id string )`)
-        .then(() => execute(session, `insert into dummy (id) values (1)`))
-        .then(() => execute(session, `
+const testComplexTypes = async (session) => {
+    try {
+        console.log('[info] create dummy');
+        await execute(session, `create table dummy( id string )`)
+        console.log('[info] insert dummy value');
+        await execute(session, `insert into dummy (id) values (1)`);
+        console.log('[info] create complexTypes');
+        await execute(session, `
             CREATE TABLE complexTypes (
                 arr_type array<string>,
                 map_type map<string, int>,
                 struct_type struct<city:string,State:string>,
                 union_type uniontype<string,double,int>
             )
-        `))
-        .then(() => execute(session, `
+        `);
+        console.log('[info] 1. insert complexTypes');
+        await execute(session, `
             INSERT INTO table complexTypes SELECT
                 array('a', 'b') as arr_type,
                 map('key', 12) as map_type,
                 named_struct('city','Tampa','State','FL') as struct_type,
                 create_union(0, 'value', cast(0.0 as double), 0) as union_type
             FROM dummy
-        `))
-        .then(() => execute(session, `
+        `);
+        console.log('[info] 2. insert complexTypes');
+        await execute(session, `
             INSERT INTO table complexTypes SELECT
                 array('c', 'd') as arr_type,
                 map('key2', 12) as map_type,
                 named_struct('city','Albany','State','NY') as struct_type,
                 create_union(1, '', cast(0.1 as double), 0) as union_type
             FROM dummy
-        `))
-        .then(() => execute(session, `
+        `);
+        console.log('[info] 3. insert complexTypes');
+        await execute(session, `
             INSERT INTO table complexTypes SELECT
                 array('e', 'd') as arr_type,
                 map('key2', 13) as map_type,
                 named_struct('city','Los Angeles','State','CA') as struct_type,
                 create_union(2, '', cast(0.0 as double), 12) as union_type
             FROM dummy
-        `))
-        .then(() => {
-            return execute(session, 'select * from complexTypes')
-        })
-        .then(console.log)
-        .then(() => Promise.all([
+        `);
+        console.log('[info] fetch complexTypes');
+        console.log(await execute(session, 'select * from complexTypes'));
+    } finally {
+        await Promise.all([
             execute(session, 'drop table dummy'),
             execute(session, 'drop table complexTypes')
-        ]), (error) => {
-            console.error(error);
-            return Promise.all([
-                execute(session, 'drop table dummy'),
-                execute(session, 'drop table complexTypes')
-            ])
-        });
+        ]);
+    }
 };
 
-const execute = (session, statement) => {
-    return session.executeStatement(statement, { runAsync: true })
-        .then((operation) => {
-            return handleOperation(operation, {
-                progress: true,
-                callback: (stateResponse) => {
-                    console.log(stateResponse.taskStatus);
-                }
-            });
-        }).then(console.log);
+const execute = async (session, statement) => {
+    const operation = await session.executeStatement(statement, { runAsync: true });
+    
+    return await handleOperation(operation, {
+        progress: true,
+        callback: (stateResponse) => {
+            if (stateResponse.taskStatus) {
+                console.log(stateResponse.taskStatus);
+            } else {
+                console.log(utils.formatProgress(stateResponse.progressUpdateResponse));
+            }
+        }
+    });
 };
 
-const handleOperation = (operation, {
+const handleOperation = async (operation, {
     progress = false,
     callback = () => {},
 }) => {
-    return utils.waitUntilReady(operation, progress, callback)
-    .then((operation) => {
-        return utils.fetchAll(operation);
-    })
-    .then(operation => {
-        return operation.close();
-    })
-    .then(() => {
-        return utils.getResult(operation).getValue();
-    })
+    await utils.waitUntilReady(operation, progress, callback)
+    await utils.fetchAll(operation);
+    await operation.close();
+    
+    return utils.getResult(operation).getValue();
 };