+++ lib/errors/OperationStateError.ts
@@ -1,3 +1,10 @@
 import HiveDriverError from "./HiveDriverError";
+import { GetOperationStatusResponse } from "../hive/Commands/GetOperationStatusCommand";
 
-export default class OperationStateError extends HiveDriverError {}
+export default class OperationStateError extends HiveDriverError {
+    public response?: GetOperationStatusResponse;
+
+    setResponse(response: GetOperationStatusResponse) {
+        this.response = response;
+    }
+}